Utils: Add some documentation about the constants used
Update boost and googletest.
Add some echos in the scripts.
Add a wait after running the instrumentation tests and before running the unit
tests in the run_tests_android.sh script since sometimes it fails for some
unknown reason (but may be because the instrumentation tests might not be
finished before it tries to start the unit tests).