Android UI: DrawView now when fails to load a scene, it finishes properly the RT
When the Ray Tracing engine couldn't load a scene, it wasn't changing the state
of the RT engine to FINISHED, so the next time the user pushed to render a new
scene it would not work. This commit fixes this issue.
Also changed the `values` arrays in the Samplers and in the Renderer to
`randomSequence` since it is more accurate name.
Improve some Javadoc.
Move `MB_IN_BYTES` constant to Constants and used it everywhere.