MobileRT: Add some javadocs, comments, a toString method, and catch exceptions
This commit continues to improve the code by fixing some alerts from the
Github Code scanning (CodeQL).
It adds some missing javadocs in the C Wrapper and also adds some comments
for readability.
It also adds an overwrite of the toString method in `ConfigSamples`, so it can
print it without using the default Objects#toString which is not readable.
Finally it also adds a try catch around the usage of Integer.parseInt methods
since they can throw a NumberFormatException without notice.