Android UI: Change RenderTask to use Kotlin coroutines
This commit changes the RenderTask implementation to use Kotlin's coroutines
instead of Android's AsyncTask since it was deprecated and it will be removed
due to possible memory leaks.
It was created the abstract class `AsyncTaskCoroutine` which is a simple
implementation of Android's AsyncTask but using Kotlin's coroutines as the
backend.
This, for now, looks enough to avoid using the deprecated AsyncTask and without
having to do a major refactor in the current's code logic.