Android UI: Isolate the call to Environment#getExternalStorageDirectory to only old Androids
This commit makes sure that the deprecated method
`Environment#getExternalStorageDirectory` to retrieve the SD card path, is only
called in the old Androids with SDK older than 19 (Android KITKAT) and when the
newer method doesn't find the SD card path.
In the rest of the cases, it just assumes that the SD card is not present and
throws an exception.
It also fixes getting the SD card path when using the SDK API 30.