JNI: Add call to `checkSystemError` in every method to check for possible errors
Ignore `EINVAL (Invalid argument)` and `EAGAIN (Resource unavailable, try again)`
cases in the `checkSystemError` Utils method, because it was found out that
there is an error in Android when allocating more than 20k bytes of memory,
and the size values are random, so its impossible, for now, to know the real
reason for it to occur.
Disable search for `gitsubmodule` in dependabot since they are not that relevant.
Update glm.