Android UI: Use lombok annotations to create the `RenderTask` builder
This commit replaces the `RenderTask` builder by the lombok annotation.
For it, it was necessary to create a configurator for the `RenderTask` called
`ConfigRenderTask`, which contains most of the necessary parameters. Not all
of them, because some were already being used by the `MainRenderer` class which
is the one that creates a `RenderTask`.
It also updates the boost dependency and the duplicated code report.