Android UI: Remove unnecessary `ExecutorService` in AsyncTaskCoroutine
This commit improves the `AsyncTaskCoroutine` by removing the
unnecessary `ExecutorService` used.
It now only uses the Kotlin coroutines in order to control the
background thread's and the UI thread's code that is executed.
It also updates the Gradle to the version with `all` instead.
Finally, it adds some custom directories in `gradle.build` file
for the sources and tests files for the IDE to recognize.