install_dependencies: Improve `install_dependencies` to support more Linux flavors and also MacOS
This commit improves the `install_dependencies` script to also support
more Linux flavors, so, at the moment it can install dependencies,
for the following Linux flavors:
* Debian
* Red Hat
* Arch
* Alpine
* Gentoo
And it also supports to install dependencies for MacOS as well, and
it makes sure to use the latest Homebrew version package manager,
so it can find packages from the new repository, since the support
for `Bintray` was removed.
The support for the new Linux flavors and MacOS was in the `Dockerfile`
and they were moved to the `install_dependencies` script, so it can be
used in Docker and native environments.