Dockerfile: Add installation of `tzdata` in Dockerfile for Debian Linux
This commit adds the installation of `tzdata` in Dockerfile for Debian
based Linux. This is done to prevent the need of user interaction
when configuring tzdata (it asks for timezone) since when installing it
with root (sudo), it always requires user interaction. Note that, in
Dockerfile, its not necessary root to install the packages, so that's
why it works there.