scripts: Add verification of errors in the bash scripts
This commit improves the bash scripts by adding some verifications
of errors in the scripts. To do that, it was added the following
command at the beginning of each script:
`set -euo pipefail;`
This makes the scripts fail when an error is detected, also it
avoids masking the errors when using piping, and finally it also
checks if there is any reference to any variable that wasn't
previously defined.