CI: Add step to pull MobileRT to the latest version and update the images
This commit adds a step in the `docker.yml` workflow to pull the
latest version of MobileRT in the docker images already available
and update them with `docker commit`.
This is done to make the docker builds faster since the gentoo
builds takes too much time (some hours) to finish.
By doing this modification, it makes it possible to run the
docker workflow for all images available again.