CI: Make all workflows cancel the previous run when a new is triggered
This commit improves the efficiency of the CI workflows by making them
cancel all the previous triggered workflows when a new one is
triggered.
This avoids having a bunch of workflows running in parallel.
Also, usually, its only wanted to run the latest pipeline
(most updated) in a merge request.