Android UI: Fix `MainActivity#getPathFromFile` method
This commit fixes the `MainActivity#getPathFromFile` method by adding
the support to detect if the loaded file is from a SD card or from
internal storage of Android.
It also improves the detection of SD card and internal storage paths
by verifying then if the current user has permissions to read and to
write files into those paths.