CI: Add caching of Android Sonar packages
This commit adds caching of Android Sonar packages because the
`sonarqube` step of gradle takes more than 40 minutes to download the
Sonar plugins.
This commit adds a cache for Android Sonar packages, so the 2nd time
a build is triggered, it doesn't have to download the Sonar plugins
alls over again.
Note that the caching only lasts 7 days and has a maximum of 5GB,
as it can be read here:
https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows#using-the-cache-action