CI: Add download of Android Sonar packages as Github artifact
This commit adds download of previous Android Sonar packages as Github
artifacts because the `sonarqube` step of gradle takes more than 40
minutes to download the Sonar plugins.
This commit adds the download of Sonar packages before trigger the
sonar, so the 2nd time a build is triggered, it doesn't have to
download the Sonar plugins via gradle all over again.