Android UI: Make Android Instrumentation Tests compatible with Android 11
This commit adds the support for Android 11 (API 30) in the Android
Instrumentation Tests.
It removes some unnecessary permission requests in the manifest files
and also it splits the permission grants in the Android Instrumentation
Tests since it was not possible to request multiple permissions at the
same time.
It also adds a new folder in the SD Card path that is compatible with
Android 11 emulator to where the files should be copied to.