Android Tests: Add call of glReleaseShaderCompiler in MainRendererTest
This commit adds a call of `glReleaseShaderCompiler` in
`MainRendererTest` test class after compiling and then deleting the
GLSL shader.
It was found out that the test `MainRendererTest` has some GPU memory
leak since if it is executed some hundred times the GPU will eventually
not be able to compile more GLSL shaders.