scripts: Kill `graphics.allocator` process before starting the tests
This commit adds a kill of the `graphics.allocator` process since it
has a bug where it accumulates a memory leak by continuously using
more and more files of '/dev/goldfish_pipe' and never freeing them.
It also sets the ADB APK installation timeout to only 3 minutes since
that should be enough for most cases.