renovate.json: Add auto merge requests of multiple types of dependencies
This commit updates the `renovate.json` to make the Renovate to
auto merge the opened merge requests of multiple types of dependencies
when all the tests pass without any issue. It should only update if
all tests pass the the update is only a minor or a patch update.
The goal is let Renovate update dependencies that should not break
anything (considering the test coverage is enough of course).