third_party: Update `CMakeLists.txt` to always rebase when pulling commits
This commit updates the `app/third_party/CMakeLists.txt` file to always
rebase instead of merging when pulling new commits from the third party
repositories.
It also makes the clean of the third party repositories to also clean
the ignored files instead of only the ignored files.
Finally it removes the recursive pulling of commits from the
dependencies of these repositories since it might bring code that is
not necessary to compile this project.
The goal of this commit is avoid conflicts as much as possible when
pulling new commits from the third party repositories.