docker_image: Fix the Dockerfile's entrypoint and default command
This commit fixes the "ENTRYPOINT" and "CMD" layers in the `Dockerfile`.
It also updates `glm` and `googletest` third party dependencies.
Before this commit, the "ENTRYPOINT" would not execute the default
command set on the Dockerfile since the "CMD" was in shell format.
This commit fixes it by making "CMD" as default parameters to the
"ENTRYPOINT.