third_party: Add missing `"` to the `git submodule foreach` commands
Summary:
This commit adds some `"` to the commands inside `git submodule foreach`
so its easier to read.
It also adds ` || true` to the foreach that performs the `git pull` to
all the git submodules, so if it fails in one, it still should perform
to the rest of the submodules.
Finally, it also updates the `glm` third party dependency.

Test plan:
* Tested compile locally: `bash scripts/compile_native.sh -t debug`
* Tested execute locally: `bash scripts/run_tests.sh -t debug`