System_dependent: Change some methods from `final` to `override`
Summary:
This commit changes some methods in `System_dependent` from using the
`final` keyword to use `override`. This is because these Qt methods
are not inside a `final` class and some compilers would give an error
because of it.
It also adds Java 11 in the workdir of `android_repeatable` workflow.

Test plan:
* Run the pipeline