scripts: Split `install_dependencies.sh` script into multiple functions
Summary:
This commit splits the script `install_dependencies.sh` into multiple
functions to install necessary dependencies on different operating
systems.
It also adds the support to execute `shellcheck` on that script
when it is available on the machine.
Finally it also splits some functions in `compile_native.sh` script so
its easier to read and maintain.

Test plan:
* Test locally