workflows: Add usage of all Operating Systems in Github workflows
Summary:
This commit adds the usage of all available Operating Systems in
runners of Github workflows.
So, from now on, these Operating Systems will be used on the Github
workflow native builds:
* Windows Server 2022
* Windows Server 2019
* Ubuntu 22.04
* Ubuntu 20.04
* Ubuntu 18.04
* macOS Monterey 12
* macOS Big Sur 11
* macOS Catalina 10.15
The goal of this change is to guarantee that MobileRT is compatible
with multiple Operating Systems and with different versions.

Test plan:
* Trigger pipeline