docs: Add gitbook badge to README
Summary:
This commit adds the gitbook badge to the README file. This new badge
will point out to the gitbook documentation URL:
https://tiago-s.gitbook.io/mobilert/docs
It also makes `codeql-analysis` github workflow to ignore any
modifications in the `docs/` path.
Also makes `github-pages` workflow to make sure to be executed if
if any modification to the `docs/` path is detected in a commit.

Test plan:
* N/A