workflows: Add more Android versions in CI/CD
Summary:
This commit adds more Android versions to be used on the tests in CI/CD.
It also refactors the way Android UI is updated with messages to the
user so its possible to show a message from any part of the code. By
doing this, its easier to let the user know why the system crashed
or why can't render a scene.
It also fixes the support for Android 4.1+ (until 4.4 inclusive at
least) by resetting the `errno` after creating the Bitmap for the
Ray Tracing preview (the rasterization of a frame before calling the
Ray Tracing engine).
Finally, it also ignores some bad file descriptor "errno" errors that
happens on Android 5+, and it also displays possible errors that might
occur when starting to render a scene and it now doesn't throw any
error if the SD card can be read but not written since it doesn't
matter on the engine to have write privileges.

Test plan:
* Trigger pipeline