chore(scripts): Improve `run_tests_android.sh` script to dynamically detect SD card path
Summary:
This commit improves the `run_tests_android.sh` script to dynamically
detect the SD card path. This is useful because each Android version
might use a different folder name to mount the SD card inside the
`/storage/` path. The goal is to prepare the scripts to allow to be
used with Android emulator with different versions besides only the
Android 4.1 and Android 11.
It also improves the `unlockDevice` method to make sure it unlocks the
Android emulator after it boots.

Test plan:
* Trigger pipeline