fix(renovate): Add property `recreateClosed` as `true`
Summary:
This commit adds the property `recreateClosed` as `true` to Renovate
config, so it also recreates pull requests even if the same ones were
closed previously.
It also makes the `renovate.json` config file to be ignored by all
github actions since it is a configuration that doesn't affect the
project itself.
Finally, it also fixes the `fileMatch` and `matchStrings` fields in the
`regexManagers` of renovate config file and also adds `ubuntu` as an
ignored dependency of `regex` manager.

Test plan:
* Run Renovate
* Check Renovate logs in:
  https://app.renovatebot.com/dashboard#github/TiagoMSSantos/MobileRT
* Validate ubuntu update was ignored for both Dockerfile and YAML files