fix(scripts): Disable Qt5 in MacOS build of `native.yml` workflow
Summary:
This commit adds the installation of `cmake` for MacOS in the
`install_dependencies.sh` script.
This is necessary otherwise the pipeline for MacOS fails with the error:
```
Command 'cmake' is NOT installed.
```
It also removes the setup of Qt5 in `native.yml` workflow since it's
currently not being installed via homebrew on the CI/CD pipeline. The
Qt4 will be used as default for MacOS builds.
Finally, also fix the created symbolic links for Qt4 headers in
`native.yml` workflow.

Test plan:
* Trigger Native (MacOS) pipeline