ci(workflows): Add missing timeouts to some of github workflows
Summary:
This commit adds some missing timeouts to the github workflows. The
goal is for the workflow to fail when it gets stuck at doing something
like: downloading dependencies, running Android tests, etc. By doing
this, the feedback will be faster if the user retries the workflow as
soon it fails due to timeout instead of having to wait more time and
still fails anyway.
Note that there is still the possibily to add more timeouts at each
workflow step.

Test plan:
* Trigger pipeline