ci(workflows): Extract downloaded `sonar-packages` artifact
Summary:
This commit adds the extraction of the downloaded `sonar-packages`
artifact in the `android.yml` workflow.
It also changes some github actions plugins from a specific patch
version to the current major being used. This is recommended in order
to avoid having to update the github action plugins too often since
the developers can update the major tag to point to the latest patch
version by themselves. Note that a plugin version is defined like
`<owner>/<plugin>@<version>` where ``<owner>/<plugin>` is a github
repository and `<version>` can be a git tag, branch or even a specific
commit reference.

Test plan:
* Trigger pipeline