ci(workflows): Remove usage of absolute path in `action-download-artifact`
Summary:
This commit removes the usage of absolute path in the github action
`action-download-artifact` to specify the directory where to extract
the artifact(s). It now uses the environment variable `${HOME}` instead
of `/home/runner` in order to make it compatible with any runner in
case the home path is different.

Test plan:
* Trigger pipeline