chore(scripts): Improve shebang portability on the bash scripts
Summary:
This commit improves the shebangs used on the bash scripts by replacing
the shebang used `#!/bin/bash` by `#!/usr/bin/env bash`.
Basically, the `/usr/bin/env` run a program such as a bash in a
modified environment. It makes the bash script portable. The advantage
of `#!/usr/bin/env bash` is that it will use whatever bash executable
appears first in the running user's $PATH variable. For more
information, please check: https://bash.cyberciti.biz/guide/Shebang
It also adds a validation for the generated files of code coverage
in the `generateCodeCoverage` bash function.
Finally, it also replaces the usage of `source` by `.` (dot) so it is
more portable as well since source is a synonym for dot/period '.' in
bash, but not in POSIX sh, so for maximum compatibility use the period.
For more info, check: https://ss64.com/bash/source.html

Test plan:
* Trigger pipeline