ci(workflows): Parallelize Android workflow
Summary:
This commit parallelizes the Android workflow.
Now there are 4 jobs:
* Build
* Unit tests
* Android tests
* Sonar
It was used caching to pass downloaded and generated artifacts between
these jobs. The cache is used for the `~/.gradle` and `.gradle` paths
where the downloaded dependencies are stored, and also used for
`app/build`, `app/.cxx` and `app/third_party` to store the generated
binaries.
It also makes the Android workflow to only build MobileRT for `x86`
Android CPUs. Since the emulator used for the tests has a CPU with
architecture `x86`.
Finally, the Android workflow now only uses the NDK version
`23.2.8568313` since it is the only version that the runners provide
out of the box and which is compatible with MobileRT currently.

Test plan:
* Trigger pipeline