ci(workflows): Add download & upload of gradle dependencies as artifacts
Summary:
This commit adds the download and upload of gradle dependencies as
github artifacts. The goal is to reduce the usage of caching for the
gradle dependencies since this can take a lot of space for a cache as
it can be a folder with size of 1GB or maybe even 2GBs.

Test plan:
* Trigger pipeline