ci(workflows): Update `ninja`, `ccache` and `cmake` in Windows workflow
Summary:
This commit updates the `ninja`, `ccache` and `cmake` versions in the
github Windows workflow. The cmake is now the version `3.24.1`, the
ninja is `1.11.0 and the ccache is the version `4.6`.
It also fixes the matrix in the Docker workflow in order for the
Renovate be able to update the docker images used.

Test plan:
* Trigger pipeline