ci(workflows): Replace the usage of `action-download-artifact` by `download-artifact`
Summary:
This commit replaces the usage of the github action
`dawidd6/action-download-artifact` by the official github action
`actions/download-artifact` for the cases where its used to pass
artifacts from a job to another job in the same workflow.
The `dawidd6/action-download-artifact` is only necessary when the goal
is to download artifacts from another workflows or from previous job
runs of the same workflow.

Test plan:
* Trigger pipeline