chore(renovate): Disable `dependencyDashboardApproval` for major updates
Summary:
This commit disables the `dependencyDashboardApproval` setting for the
major updates. This setting means that is necessary for the user to
approve first before Renovate opens a pull request, which is
unnecessary. We just want the user to have to review the pull request
before merging which Renovate is already configured for that.

Test plan:
* Trigger pipeline