build(gitmodules): Change `googletest` branch to `main` in `.gitmodules`
Summary:
This commit changes the `googletest` git submodule branch to `main` in
the `.gitmodules` file.
It also updates the `googletest` and `glm` git submodules.
Finally, it also changes Xcode to version `13.2` in MacOS builds
because version `14.0.1` has a bug where MobileRT can't be properly
linked. And `13.2` is the only version available for both MacOS-11 and
MacOS-12,

Test plan:
* Trigger pipeline