refactor(scripts): Make shell scripts POSIX compliant
Summary:
This commit makes all of the shell scripts, used by MobileRT, POSIX
compliant.
This change improves the compatibility of those scripts by allowing
any Operating System which implements the POSIX interface to
automatically be compatible with the scripts.
It also changes some of the shell scripts to be used by shell instead
of bash so it's even more compatible with Operating Systems without
having to install the `bash`. It's not possible to change all of the
scripts to shell because some of them use the `gradlew` script which
needs `bash` and also because some use arrays in the script which is
an exclusive feature of bash scripts and not supported by shell.

Test plan:
* Trigger pipeline