fix(deps): update dependency org.sonarsource.scanner.gradle:sonarqube-gradle-plugin to v3.5.0.2730
This commit updates the dependency
`org.sonarsource.scanner.gradle:sonarqube-gradle-plugin` to the version
`v3.5.0.2730`.
It also fixes the function `clearOldBuildFiles` in the
`helper_functions.sh` script.
Finally, it also makes `MobileRT` and `Components` C++ libraries
accessible when linking the `AppMobileRT` in Java including all their
dependencies like stdlib, otherwise the Android tests would fail
with NDK `24.0.8215888` stating that the functions in C++ library were
not linked to the project. And it also links the `android` and
`jnigraphics` C libraries to the Unit Tests app because of similar
issue.