test(scripts): Add some unit tests for the shell scripts
Summary:
This commit adds some unit tests for the functions in the shell
scripts. For now, it only adds some unit tests to the functions of
the `helper_functions.sh` script. The goal is for later commits to
continue add more unit tests and have most of the shell scripts
functions being tested in CI/CD.
These shell unit tests will be executed in the new step
`Shell scripts unit tests` of the `Code Analysis` github workflow
(which is the old `CodeQL` workflow).

Test plan:
* Trigger pipeline