refactor(scripts): Replace " by ' when possible
Summary:
This commit replaces the usage of double quotes (") by single quotes
(') whenever possible in the shell scripts. The goal is to easily allow
the reader of the code to understand when the echos used do need to
expand the argument or not. If the single quotes is used, then it's
easy to understand that the text written as argument of the `echo` does
not need to expand anything.

Test plan:
* Trigger pipeline