refactor(scripts): Make all the shell scripts not need to use bash
Summary:
This commit makes all the shell scripts to not need to use bash. The
goal is to make sure MobileRT is compatible with Operating Systems that
don't have bash installed by default and need to use shell instead. So
bash does not need to be a requirement to execute MobileRT.

Test plan:
* Trigger pipeline