test(scripts): Add more unit tests
Summary:
This commit adds more unit tests to the functions in the shell scripts.
It also replaces the usage of double quotes (") by single quotes (')
when possible, since they should be used when there is not anything in
the shell string that could be expanded (e.g.: call to shell functions,
or usage of shell environment variables).

Test plan:
* Trigger pipeline