fix(scripts): Make `yum install` commands validate if found dependencies
Summary:
This commits makes all the `yum install` commands validate if every
dependency ordered to be installed or update, were really
installed/updated or if some were not found. If some dependency was not
found, then the CI build will fail.
The `yum install` flags added are:
* --setopt=skip_missing_names_on_install=False to validate if every
dependency was installed as requested
* --setopt=skip_missing_names_on_update=False to validate if every
dependency was updated as requested
For more information, please check:
https://unix.stackexchange.com/questions/165326/is-there-a-yum-option-that-creates-an-error-if-a-package-is-missing
https://dashdash.io/5/yum.conf

Test plan:
* Trigger pipeline