fix(app): Merge Jacoco Unit Tests & Instrumentation Tests reports
Summary:
This commit fixes the Jacoco report bv merging both the reports
generated by the Unit Tests with the report generated by the Android
Instrumentation Tests. By doing this, the new merged report has a more
accurate code coverage percentage.
It also adds the support for the PowerMock in the Unit Tests, which is
useful to mock native methods and static methods, which MobileRT java
interface has a lot.
Finally it also adds the `MainActivityTest` unit test class which tests
the `MainActivity` class. These new unit tests use Mockito and PowerMock
to mock the necessary methods from the Android API, in order to test
the implemented code.

Test plan:
* Trigger pipeline