test(mobilertapp): Add more unit tests for `MainRendererTest` test class
Summary:
This commit adds more unit tests for the `MainRendererTest` test class.
The goal is to increase the code coverage as much as possible, because
the Android Instrumentation Tests do not cover every case, specially
the cases where something might fail during runtime.
It also makes Renovate ignore the update of these Gradle dependencies:
* `androidx.test:rules`
* `androidx.test:runner`
Because the update can't be done due to the Android Tests failing in CI
and Renovate keeps opening the same Pull Requests to update them, which
shouldn't happen due to the property `recreateClosed` is already set
as `false`.
Finally it also adds the support for lombok Android Studio plugin for
the tests. And note that this plugin is compatible with Android Studio
2021.3.1 Patch 1, as it can be seen here:
https://plugins.jetbrains.com/plugin/6317-lombok/reviews#review=64000
https://github.com/mplushnikov/lombok-intellij-plugin/issues/1079

Test plan:
* Trigger pipeline