test(mobilertapp): Add `DrawViewTest#testRenderSceneWithLowMemory` test
Summary:
This commit adds the Java unit test
`DrawViewTest#testRenderSceneWithLowMemory` which validates that the
user is notified when there is not enough memory available to render
a scene.
It also adds the EasyMock dependency to be able to create mocks
and suppress native methods in the mocked class. And adds the
`spring-test` dependency to be able to set private fields in any mocked
class from the EasyMock framework with the `ReflectionTestUtils` utils
class.

Test plan:
* Trigger pipeline