fix(deps): Revert "update kotlin monorepo to v1.8.0"
This reverts commit 8c1104e6e6ab61664014c3f6255b1bca21139258.
Because CodeQL is not compatible with Kotlin 1.8:
```
Kotlin version 1.8.0 is too recent. CodeQL currently supports versions below 1.7.30
```
It also makes the Code Analysis github workflow to be executed in every
commit that modifies build.gradle files since it might mean doing an
update of a dependency, and thus Code Analysis should be executed when
that happens.