test(mobilertapp): Add `PreviewTest#testPreviewSceneOrthographicCamera`
Summary:
This commit adds the `PreviewTest#testPreviewSceneOrthographicCamera`
Android Instrumentation test which validates that the preview feature
works for scenes with orthographic camera.
It also removes the unnecessary `Scene#WRONG_FILE` which was an enum
value used only for the tests to validate the behaviour of MobileRT
why trying to load a file that doesn't exist. It is now used the
`Intents#intending` method to mock the reply from an external file
manager application by pointing it to a file that doesn't exist.

Test plan:
* Trigger pipeline