fix(mobilertapp): Fix detection of SD card path
Summary:
This commit improves the `MainActivity#getPathFromFile` method by
making it also detect paths starting with `/sdcard` as an external SD
card path.
It also makes the Native github workflow to be executed when the Android
github workflow is executed, so the code coverage is always correct with
both Android and Native code being analyzed in the report.

Test plan:
* Trigger pipeline