fix(MobileRT): Support Android 4.0.3 (API 15)
Summary:
This commit makes MobileRT compatible with Android 4.0.3 (API 15). It
does small refactors because of it and also makes Android 4.0.3 to be
tested in CI. And updates the README accordingly.
In reality, the minimum Android SDK required is Android API 14 (or
Android 4.0) due to compatibility with Appcompat dependency but it was
never tested because there is no Android Emulator X86 with Android 4.0.

Test plan:
* Trigger pipeline