build(MobileRT): Prepare support for Android API 33
Summary:
This commit adds prepares support for Android API 33.
It refactors the code in order to allow the Android interface to perform new
Intents in order to read the necessary file descriptions for the scene files
and then MobileRT can read the files by those file descriptions.
Unfortunately it still doesn't allow to read the scenes files from an external
SD Card because it requires different permissions that Android API 32 and older
requires, which is just `READ_EXTERNAL_STORAGE`.
Android 33 and newer requires the following permissions to read files from SD card:
* READ_MEDIA_IMAGES
* READ_MEDIA_VIDEO
* READ_MEDIA_AUDIO
But after adding those permissions it still doesn't work properly because it
outputs the following error when asking for permission to read the files:
```
GrantPermissionsViewModel: None of [android.permission.READ_MEDIA_IMAGES] in {}
```

Test plan:
* Trigger pipeline