fix(build.gradle): Enforce Kotlin version
This commit enforces the version of 'org.jetbrains.kotlin:kotlin-stdlib'
dependency to be 1.8.20, because the dependency 'androidx.core:core-ktx'
also brings it as transitive dependency.
It also improves the Linux Gentoo build by adding some echos to understand
better when a ci pipeline fails, and also avoid using `pip` directly, and
use it within `python` command instead, so it works with Gentoo.