test(mobilertapp): Add RayTracing tests which use every shader available
This commits add some tests to the `RayTracingTest` test class which tests
the Ray Tracing engine with every shader available.
These new tests also use the scene CornellBox because it only has a point light
and thus it should never be rendered in the scene like the area lights do. This
is important in order to validate that the primitives in the scene were properly
rendered as expected without taking the light itself into account.