ci(codecov): Increase max report age to 24 hours
This commit increases the maximum report age of CodeCov to 24 hours
because sometimes an Android test might fail during the night
and if someone retries in the next day, the new code coverage for
Android tests might not update the CodeCov report due to timeout.
With this commit, it should now be possible to update a code coverage
report of CodeCov if something similar happens again.