chore(mobilertapp): Migrate RenderTask to Kotlin
This commit migrates the class ´RenderTask´ to Kotlin since it is the official
Android language now. It also removes completely the usage of lombok since the
Android IDE does not recognize many annotations of it.
Finally it also reduces the Android emulator timeout in CI to 6 minutes since
it usually boots with less than 5 minutes.