test(MobileRT): Add some basic ray tracing tests as unit tests
This commits adds some basic ray tracing tests, which uses the whole
Ray Tracing engine with different shaders and different accelerators.
These new unit tests are a copy of Android tests that are used to
have a more accurate code coverage report because Jacoco has a limitation
which doesn't support code coverage of native code. So, a workaround for that
is duplicating the same Android tests as native unit tests and thus it uses
LCOV instead. Also note that these new tests do not validate anything, they
only serve to make sure the engine is properly exercised and code coverage
is updated like it should be by the Android Instrumentation tests.