ci(engine tests): Make all CI pipelines run the Ray Tracing Engine tests
This commit makes all the CI pipelines to also run the C++ Ray Tracing Engine tests
instead of only native debug pipeline execute it, so it takes advantage of executing
the tests with different operating systems and with different compilers.
It also adds some assertions after executing the Ray Tracing Engine tests to
validate that the values in the Bitmap were changed and that they are not equal to the
same value (a very simple validation of Bitmap output).
It also fixes the output result of native unit tests in Android emulator,
and found out that those unit tests are failing for Android API 15. So, the
result of unit tests is ignored for Android API 15 for now.
And finally, it was done the command `git init` to fix the permissions of all
files in the repository.