fix(scripts): Fix copy of Unit tests + libs to Android x64
Before this commit, the Unit Tests and respective necessary libs that were
copied to Android emulator was always the binaries and libs for CPU x86
architecture and thus the Uni Tests always failed when executed on
Android x64. This commit fixes this issue by proper choosing the binaries
to copy to the Android emulator.