chore(scripts): Unlock device before install APKs
Also reduce the RAM used by Android emulator from 8GB to 1GB since it was
detected that sometimes the tests would fail with an "Out of memory" error
from the C++ native code.
Also added some verifications of `errno` errors at the start and end of the
`Renderer#renderScene` method by every thread.
And finally made the JNI layer to always call the `DetachCurrentThread` function
when the render of the scene ends because it was also detected that sometimes
the tests would fail when a Renderer thread finished its work and stated
that the thread exited without performing the detach, which could mean
that a test could fail due to a previous test didn't call the detach function.