fix(mobilertapp): Remove usage of weak references in `RenderTask`
It also improves the Android instrumentation tests by making sure that when the
method `Espresso#onIdle` is called, it really waits for the app to become idle.
Finally, it also detaches the JVM thread while the Ray Tracing engine is working
to avoid any possible error that happened when the Renderer threads finished their
work stating that the thread was finished but not detached, and then reattaches
the JVM thread after the Renderer is finished.