refactor(mobilertapp): Replace javax Nonnull by Android X NonNull
It also replaces the javax Nullable by Android X Nullable. This was done so all
the annotations are from the same package.
It also deletes the `OverridingMethodsMustInvokeSuper` annotation from javax
because the `CallSuper` is already being used from Android X package.
And it deletes the `UtilsT#invokePrivateMethod` method because it was not being
used.
And it removes the dependencies `com.github.spotbugs:spotbugs-annotations`,
`org.apache.commons:commons-lang3`, `com.android.volley:volley` and `com.fasterxml.jackson.core`
from gradle because they were not being used.
Also note that the method `ViewInteraction#check` should only be called when the
app is idle, otherwise it can throw an exception.