chore(renovate): Ignore `pcg-cpp` conan dependency
This commit makes Renovate ignore the `pcg-cpp` conan dependency.
It is done because Renovate can't properly determine the more recent versions
of conan dependencies like `pcg-cpp` that are based on dates.
For example, the pull request https://github.com/TiagoMSSantos/MobileRT/pull/574
that was opened automatically by Renovate tried to update `pcg-cpp` from version
`cci.20220409` to `cci.20210406` which is an outdated release.