ci(workflows): Make CodeQL workflow also inspect Ruby code
This commit enables the CodeQL workflow to also inspect the Ruby code found in
the repository, namely the Gemfile.
It removes the usage of the secret `REPOSITORY_TOKEN` because the token was
deprecated and github already provides the secret `GITHUB_TOKEN` with the
current token being used by the current github account anyway.
Finally it also makes the workflow `Sync GitHub Page` being triggered by
`gh-pages` branch. This change is useful in order to use that branch when
updating documentation and having a way to render it in CI without having to
merge the changes in `master` branch.