chore(mobilertapp): Add validation of file paths selected by user
This commit makes MobileRT app to only accept paths of files that are in an
external SD card or in path `/data/local/tmp` of local storage.
The goal is to not allow the user to select paths that might be unsecure
because they can contain some important files for the Operating System.