fix(workflows): Ignore execute `Autobuild` step for C++ build
This is necessary because otherwise CodeQL will output the following error
`All detected build commands failed (tried make).`.
The next next `Build C++` step will build the project for the native code
which CodeQL build was already using as 2nd attempt of building the project
and will now be only one in order to avoid the error reported above.