ci(native): Add some tests to profile script
This commit adds a basic test of the profile shell script to the native workflow.
The new test uses 'timeout' command to validate whether the profile script boots
and starts to render the scene. It only tests the 'release' and 'debug' modes
of the script for now.
The goal is to detect if changes to the profile script breaks it or not.