fix(mobilertapp): Remove asking for reading SD card when using Android API 34+
This is necessary for Android API 34 because it seems that if MobileRT asks for
permissions, then Android will reset all permissions after doing so.
Now the necessary permissions are given to MobileRT and MobileRT_test APKs
before starting to execute the Android instrumentation tests in
`run_tests_android` shell script.
It also updates the following Android API tested in CI:
* Replaced Android API 33 to Android API 34